+++ data-structures/blockchain-ghost/ghost-transaction.go
@@ -1,0 +1,10 @@
+package ghost
+
+// What a transaction ensues
+// A transaction is a request to move $X from A to B
+type transaction struct {
+	origin          string
+	senderSignature string
+	destination     string
+	value           float32
+}
