+++ examples/blockchain/blockchain.go
@@ -1,0 +1,1 @@
+package main
